Gemma OCR App/requirements.txt
streamlit
streamlit-extras
ollama
pillow
pyperclip